backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.39.0
chromadb==0.5.0
sentence-transformers==3.0.0
pydantic==2.9.0
python-dotenv==1.0.1
langdetect==1.0.9
httpx==0.27.0
